Change package structure
fixup! Add more structure and a `query`-backend
