+++ .gitignore
@@ -3,9 +3,9 @@ __pycache__/
 .pytest_cache/
 logs/
 .DS_Store
+tests/data/
 
 # Specific files to ignore
 
 # Folders to ignore
 other_scripts/
-tests/
